adjust tokenvalidation library and remove old auth code from transfer
