+++ domains/libraries/EnergyOrigin.TokenValidation/EnergyOrigin.TokenValidation/B2C/ServiceCollectionExtensions.cs
@@ -1,6 +1,3 @@
-using EnergyOrigin.TokenValidation.Options;
-using EnergyOrigin.TokenValidation.Utilities;
-using EnergyOrigin.TokenValidation.Values;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authorization.Infrastructure;
 using Microsoft.Extensions.DependencyInjection;
@@ -13,11 +10,8 @@ public static class ServiceCollectionExtensions
 
     public static readonly List<string> BooleanTrueClaimValues = ["true", "True"];
 
-    public static void AddB2CAndTokenValidation(this IServiceCollection services, B2COptions b2COptions, TokenValidationOptions validationOptions)
+    public static void AddB2C(this IServiceCollection services, B2COptions b2COptions)
     {
-        var tokenValidationParameters = new ValidationParameters(validationOptions.PublicKey);
-        tokenValidationParameters.ValidIssuer = validationOptions.Issuer;
-        tokenValidationParameters.ValidAudience = validationOptions.Audience;
 
         services.AddAuthentication(defaultScheme: AuthenticationScheme.TokenValidation)
             .AddJwtBearer(AuthenticationScheme.B2CAuthenticationScheme, options =>
@@ -40,11 +34,6 @@ public static class ServiceCollectionExtensions
                 options.TokenValidationParameters.ValidateAudience = false;
                 options.TokenValidationParameters.AudienceValidator = (_, _, _) => true;
                 options.MetadataAddress = b2COptions.MitIDCustomPolicyWellKnownUrl;
-            })
-            .AddJwtBearer(AuthenticationScheme.TokenValidation, options =>
-            {
-                options.MapInboundClaims = false;
-                options.TokenValidationParameters = tokenValidationParameters;
             });
 
         services.AddAuthorization(options =>
@@ -81,14 +70,6 @@ public static class ServiceCollectionExtensions
                 .AddRequirements(new ClaimsAuthorizationRequirement(ClaimType.Sub, new List<string> { b2COptions.CustomPolicyClientId }))
                 .Build();
             options.AddPolicy(Policy.B2CInternal, b2CInternalPolicy);
-
-            // TODO: Cleanup after V1 release
-            var tokenValidationRequiredCompanyPolicy = new AuthorizationPolicyBuilder()
-                .RequireAuthenticatedUser()
-                .AddAuthenticationSchemes(AuthenticationScheme.TokenValidation)
-                .RequireClaim(UserClaimName.Tin)
-                .Build();
-            options.AddPolicy(PolicyName.RequiresCompany, tokenValidationRequiredCompanyPolicy);
         });
 
         services.AddScoped<IdentityDescriptor>();
