+++ domains/transfer/ClaimAutomation/Worker/appsettings.Development.json
@@ -17,8 +17,5 @@
     },
     "Otlp": {
         "ReceiverEndpoint": "http://localhost:4317"
-    },
-    "TokenValidation:Issuer": "demo.energioprindelse.dk",
-    "TokenValidation:Audience": "Users",
-    "TokenValidation:PublicKey": "LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUlJQklqQU5CZ2txaGtpRzl3MEJBUUVGQUFPQ0FROEFNSUlCQ2dLQ0FRRUFvc1IxV1lKODYvWDVmVXRMdHllWQpjcU5HeHdDd21yd3FtZ29aRWFKQ3FXcUVDTVMydlp2U1kwajNvazlBdDFqckVEeHd0cnEzbUFid1NHbi9ycDlrCjRTV3JmSjVUOGhWVHdlS0IweHZoUStZSWh1akdVMW10NEIvai9ScC9hSXZ2cHhiQ01QZzdaVVAyK3pYQm9QaUkKUGxpS3RoQnQ4V3VnQ1YreTBCdW9ENnAyUmI2SlowZTlBS0MzeHlzN1N5TGpBb2RoZWxoZWFlVUFlQXJyWmxUcApNd01nV1dOQTRiK2FjSkxRVnM4QVBBdmpxcG02SHZnS3dkVlUwQzk3SmV6ZnF4cW9jOUpjT1VzL296cGVMQlBpCmtwTllnYnkwV2tsS0l2V2RJTzJGdmRXSzdybnpNZmg3SXBLM0lJQmRNOCtESmFINEdaVHI1TjVUN1VYclorSVkKOHdJREFRQUIKLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg=="
+    }
 }
