--- domains/transfer/Transfer.API/API.IntegrationTests/Cvr/Api/Controllers/CvrController20240515Tests.GetCvrCompanies_WhenCorrectCvrNumbers_ShouldReturnCvrInfo.verified.txt
@@ -1,21 +1,0 @@
-﻿{
-  Result: [
-    {
-      CompanyCvr: 10150817,
-      CompanyName: Erhvervsstyrelsen,
-      Address: {
-        Landekode: DK,
-        Vejkode: 4143,
-        Kommune: {
-          KommuneKode: 101,
-          KommuneNavn: KØBENHAVN
-        },
-        HusnummerFra: 17,
-        AdresseId: 9ccc4f1d-2470-4f2f-a40e-17daacabd038,
-        Vejnavn: Langelinie Allé,
-        Postnummer: 2100,
-        Postdistrikt: København Ø
-      }
-    }
-  ]
-}
