+++ domains/transfer/Transfer.API/API.IntegrationTests/Cvr/Api/Controllers/CvrEmptyTests.cs
@@ -41,8 +41,7 @@ public class CvrEmptyTests
                 .WithBodyFromFile("Cvr/Api/Controllers/CvrControllerTests.empty_cvr_response.json")
             );
 
-        using var client = factory.CreateAuthenticatedClient(sub: Guid.NewGuid().ToString(),
-            apiVersion: ApiVersions.Version20240103);
+        using var client = factory.CreateB2CAuthenticatedClient(Guid.NewGuid(), Guid.NewGuid());
 
         using var response = await client.PostAsJsonAsync("api/transfer/cvr", cvrNumbers);
