+++ domains/transfer/Transfer.API/API.IntegrationTests/Cvr/Api/Controllers/CvrMultipleResponseTests.cs
@@ -42,8 +42,8 @@ public class CvrMultipleResponseTests
                 .WithBodyFromFile("Cvr/Api/Controllers/CvrControllerTests.cvr_multiple_companies_response.json")
             );
 
-        using var client = factory.CreateAuthenticatedClient(sub: Guid.NewGuid().ToString(),
-            apiVersion: ApiVersions.Version20240103);
+        using var client = factory.CreateB2CAuthenticatedClient(Guid.NewGuid(), Guid.NewGuid());
+
 
         using var response = await client.PostAsJsonAsync("api/transfer/cvr", wrongPlusRightCvr);
