+++ domains/transfer/Transfer.API/API.IntegrationTests/Factories/TransferAgreementsApiWebApplicationFactory.cs
@@ -5,12 +5,10 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Security.Cryptography;
-using System.Text;
 using System.Threading.Tasks;
 using API.Transfer.Api.Controllers;
-using Asp.Versioning.ApiExplorer;
 using API.Transfer.TransferAgreementProposalCleanup;
+using Asp.Versioning.ApiExplorer;
 using DataContext;
 using EnergyOrigin.ActivityLog;
 using EnergyOrigin.ActivityLog.HostedService;
@@ -25,8 +23,8 @@ using Microsoft.AspNetCore.TestHost;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using ProjectOriginClients;
 using NSubstitute;
+using ProjectOriginClients;
 using AuthenticationScheme = EnergyOrigin.TokenValidation.b2c.AuthenticationScheme;
 
 namespace API.IntegrationTests.Factories;
@@ -36,8 +34,6 @@ public class TransferAgreementsApiWebApplicationFactory : WebApplicationFactory<
     public string ConnectionString { get; set; } = "";
 
     private string WalletUrl { get; set; } = "http://foo";
-
-    private byte[] PrivateKey { get; set; } = RsaKeyGenerator.GenerateTestKey();
 
     private byte[] B2CDummyPrivateKey { get; set; } = RsaKeyGenerator.GenerateTestKey();
 
@@ -58,17 +54,6 @@ public class TransferAgreementsApiWebApplicationFactory : WebApplicationFactory<
 
     protected override void ConfigureWebHost(IWebHostBuilder builder)
     {
-        var privateKeyPem = Encoding.UTF8.GetString(PrivateKey);
-        string publicKeyPem;
-
-        using (RSA rsa = RSA.Create())
-        {
-            rsa.ImportFromPem(privateKeyPem);
-            publicKeyPem = rsa.ExportRSAPublicKeyPem();
-        }
-
-        var publicKeyBase64 = Convert.ToBase64String(Encoding.UTF8.GetBytes(publicKeyPem));
-
         builder.UseSetting("Otlp:ReceiverEndpoint", OtlpReceiverEndpoint);
         builder.UseSetting("TransferAgreementProposalCleanupService:SleepTime", "00:00:03");
         builder.UseSetting("TransferAgreementCleanup:SleepTime", "00:00:03");
@@ -76,10 +61,6 @@ public class TransferAgreementsApiWebApplicationFactory : WebApplicationFactory<
         builder.UseSetting("Cvr:User", CvrUser);
         builder.UseSetting("Cvr:Password", CvrPassword);
         builder.UseSetting("ProjectOrigin:WalletUrl", WalletUrl);
-
-        builder.UseSetting("TokenValidation:PublicKey", publicKeyBase64);
-        builder.UseSetting("TokenValidation:Issuer", "demo.energioprindelse.dk");
-        builder.UseSetting("TokenValidation:Audience", "Users");
 
         builder.UseSetting("B2C:B2CWellKnownUrl",
             "https://login.microsoftonline.com/d3803538-de83-47f3-bc72-54843a8592f2/v2.0/.well-known/openid-configuration");
@@ -165,15 +146,26 @@ public class TransferAgreementsApiWebApplicationFactory : WebApplicationFactory<
     public HttpClient CreateUnauthenticatedClient()
     {
         var client = CreateClient();
-        client.DefaultRequestHeaders.Add("X-API-Version", "20240103");
+        client.DefaultRequestHeaders.Add("X-API-Version", ApiVersions.Version20240515);
         return client;
     }
 
     public HttpClient CreateAuthenticatedClient(string sub, string tin = "11223344", string name = "Peter Producent",
-        string actor = "d4f32241-442c-4043-8795-a4e6bf574e7f", string cpn = "Producent A/S", string apiVersion = "20240103")
+        string actor = "d4f32241-442c-4043-8795-a4e6bf574e7f", string cpn = "Producent A/S", string apiVersion = ApiVersions.Version20240515)
     {
         var client = CreateClient();
         AuthenticateHttpClient(client, sub: sub, tin: tin, name, actor, cpn, apiVersion: apiVersion);
+        return client;
+    }
+
+
+    private HttpClient AuthenticateHttpClient(HttpClient client, string sub, string tin = "11223344", string name = "Peter Producent",
+        string actor = "d4f32241-442c-4043-8795-a4e6bf574e7f", string cpn = "Producent A/S", string apiVersion = ApiVersions.Version20240515)
+    {
+        client.DefaultRequestHeaders.Authorization =
+            new AuthenticationHeaderValue("Bearer", GenerateB2CDummyToken(sub: sub, tin: tin, name: name, cpn: cpn));
+        client.DefaultRequestHeaders.Add("X-API-Version", apiVersion);
+
         return client;
     }
 
@@ -184,61 +176,9 @@ public class TransferAgreementsApiWebApplicationFactory : WebApplicationFactory<
         client.DefaultRequestHeaders.Authorization =
             new AuthenticationHeaderValue("Bearer",
                 GenerateB2CDummyToken(sub: sub.ToString(), tin: tin, name: name, orgId: orgId.ToString(), termsAccepted: termsAccepted));
-        client.DefaultRequestHeaders.Add("X-API-Version", apiVersion);
+        client.DefaultRequestHeaders.Add("X-API-Version", ApiVersions.Version20240515);
 
         return client;
-    }
-
-    private HttpClient AuthenticateHttpClient(HttpClient client, string sub, string tin = "11223344", string name = "Peter Producent",
-        string actor = "d4f32241-442c-4043-8795-a4e6bf574e7f", string cpn = "Producent A/S", string apiVersion = "20240103")
-    {
-        client.DefaultRequestHeaders.Authorization =
-            new AuthenticationHeaderValue("Bearer", GenerateToken(sub: sub, tin: tin, name: name, actor: actor, cpn: cpn));
-        client.DefaultRequestHeaders.Add("X-API-Version", apiVersion);
-
-        return client;
-    }
-
-    private string GenerateToken(
-        string scope = "",
-        string actor = "d4f32241-442c-4043-8795-a4e6bf574e7f",
-        string sub = "03bad0af-caeb-46e8-809c-1d35a5863bc7",
-        string tin = "11223344",
-        string cpn = "Producent A/S",
-        string name = "Peter Producent",
-        string issuer = "demo.energioprindelse.dk",
-        string audience = "Users")
-    {
-        var claims = new Dictionary<string, object>()
-        {
-            { UserClaimName.Scope, scope },
-            { UserClaimName.ActorLegacy, actor },
-            { UserClaimName.Actor, actor },
-            { UserClaimName.Tin, tin },
-            { UserClaimName.OrganizationName, cpn },
-            { JwtRegisteredClaimNames.Name, name },
-            { UserClaimName.ProviderType, ProviderType.MitIdProfessional.ToString() },
-            { UserClaimName.AllowCprLookup, "false" },
-            { UserClaimName.AccessToken, "" },
-            { UserClaimName.IdentityToken, "" },
-            { UserClaimName.ProviderKeys, "" },
-            { UserClaimName.OrganizationId, sub },
-            { UserClaimName.MatchedRoles, "" },
-            { UserClaimName.Roles, "" },
-            { UserClaimName.AssignedRoles, "" }
-        };
-
-        var signedJwtToken = new TokenSigner(PrivateKey).Sign(
-            sub,
-            name,
-            issuer,
-            audience,
-            null,
-            60,
-            claims
-        );
-
-        return signedJwtToken;
     }
 
     private string GenerateB2CDummyToken(
