+++ domains/transfer/Transfer.API/API.IntegrationTests/Transfer/Api/TestData.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using DataContext;
