+++ domains/transfer/Transfer.API/API/appsettings.Development.json
@@ -29,11 +29,6 @@
         "User": "SomeUser",
         "Password": "SomePassword"
     },
-    "TokenValidation": {
-        "Issuer": "demo.energytrackandtrace.dk",
-        "Audience": "Users",
-        "PublicKey": "LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUlJQklqQU5CZ2txaGtpRzl3MEJBUUVGQUFPQ0FROEFNSUlCQ2dLQ0FRRUFvc1IxV1lKODYvWDVmVXRMdHllWQpjcU5HeHdDd21yd3FtZ29aRWFKQ3FXcUVDTVMydlp2U1kwajNvazlBdDFqckVEeHd0cnEzbUFid1NHbi9ycDlrCjRTV3JmSjVUOGhWVHdlS0IweHZoUStZSWh1akdVMW10NEIvai9ScC9hSXZ2cHhiQ01QZzdaVVAyK3pYQm9QaUkKUGxpS3RoQnQ4V3VnQ1YreTBCdW9ENnAyUmI2SlowZTlBS0MzeHlzN1N5TGpBb2RoZWxoZWFlVUFlQXJyWmxUcApNd01nV1dOQTRiK2FjSkxRVnM4QVBBdmpxcG02SHZnS3dkVlUwQzk3SmV6ZnF4cW9jOUpjT1VzL296cGVMQlBpCmtwTllnYnkwV2tsS0l2V2RJTzJGdmRXSzdybnpNZmg3SXBLM0lJQmRNOCtESmFINEdaVHI1TjVUN1VYclorSVkKOHdJREFRQUIKLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg=="
-    },
     "B2C": {
         "B2CWellKnownUrl": "https://login.microsoftonline.com/d3803538-de83-47f3-bc72-54843a8592f2/v2.0/.well-known/openid-configuration",
         "ClientCredentialsCustomPolicyWellKnownUrl": "https://datahubeouenerginet.b2clogin.com/datahubeouenerginet.onmicrosoft.com/v2.0/.well-known/openid-configuration?p=B2C_1A_CLIENTCREDENTIALS",
